[backend] Added TaskInput and TaskOutput
When I receive a Task from the external world, this task has no id.
The id is added by MySQL while storing the task.

But when I output a task to the external world, I need to append the id,
to be able to identify the task.

Notice that I had to remove the default values in the dataclasses. This is
because dataclasses do not allow to have arguments without default values
after arguments with the default value. Since TaskOutput inherits from
TaskInput, it adds the argument "id" without default, to the end of the
args of TaskInput, which have a default value.

Removing the default value is not a problem, the important thing is to have
default values in the GraphQL schema, because those are the ones exposed to
the external world.