[backend] Changed DAO Status to string
As an enumeration it caused a bug in getAllTasks and getTaskByID. There
I retrieve the status as a string from mysql and use it idirectly to
build the TaskOutput object. But if Taskoutput stores the status as an
enum, I need a converter that converts the string to the enum, which I
think overcomplicates the code. It is enough to have an enum in the QL
type, which is exposed to the outside world, but for the database a
plain string is enough.