[backend] Factored out optional fields of graphQL
This is like what I have previously done for the internal tasks. In a
future commit I will add the type TaskUpdate and there I want the same
fields as for TaskInput, but all optionals. Since the optional fields
are all factored out, I can simply inherit from them and add the title
and status as optional fields.