[backend] Changed start/end time format
Before I had the time exposed as a Python datetime and as a timestamp in
the db. This caused problems because the two formats are inconistent and
also because the HTML returns the time as a string and not as a Python
timestamp.

Now I use TIME as db format, which stores an elapsed time. The API
exposes the time as a string with a format "hh:mm".