[backend] Fixed bug in time conversion
If the task has no end_time or start_time, the respective field is set
to None, but it is still present in the dict. So, to know if I have to
convert a end/start time, I cannot simply check if end_time/start_time
is in the dict (it will always be), but I need to check if its value is
not None.